Download link based on installation type
